+++ src/main/java/com/innowise/orderanalysis/entity/Category.java
@@ -1,4 +1,4 @@
-package com.innowise.salesAndCustomerAnalytics.entities;
+package com.innowise.orderanalysis.entity;
 
 public enum Category {
     ELECTRONICS, CLOTHING, BOOKS, HOME, BEAUTY, TOYS
