+++ src/main/java/com/innowise/orderanalysis/entity/Customer.java
@@ -1,4 +1,4 @@
-package com.innowise.salesAndCustomerAnalytics.entities;
+package com.innowise.orderanalysis.entity;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
