+++ src/main/java/com/innowise/orderanalysis/entity/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.innowise.salesAndCustomerAnalytics.entities;
+package com.innowise.orderanalysis.entity;
 
 public enum OrderStatus {
     NEW, PROCESSING, SHIPPED, DELIVERED, CANCELLED
